# openapi.yaml
openapi: 3.1.0
info:
  title: Talabalar Ro'yhati API
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
...
  /api/v1/auth/login:
    post:
      summary: Admin login
      tags:
        - Admin Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  token_type: { type: string, example: Bearer }
  /api/v1/auth/change-password:
    post:
      summary: Change admin password
      description: Change the password for the currently authenticated admin
      tags:
        - Admin Auth
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                  description: Current password
                  example: "admin123"
                newPassword:
                  type: string
                  minLength: 6
                  description: New password (minimum 6 characters)
                  example: "newpassword123"
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password changed successfully"
        '400':
          description: Validation error
        '401':
          description: Current password incorrect or unauthorized
  /api/v1/auth/profile:
    get:
      summary: Get admin profile
      description: Get the profile information of the currently authenticated admin
      tags:
        - Admin Auth
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Admin profile information
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  username:
                    type: string
                  email:
                    type: string
                  full_name:
                    type: string
                  phone:
                    type: string
                  is_active:
                    type: boolean
                  last_login_at:
                    type: string
                    format: date-time
                  created_at:
                    type: string
                    format: date-time
        '401':
          description: Unauthorized
        '404':
          description: Admin not found
    patch:
      summary: Update admin profile
      description: Update the profile information of the currently authenticated admin
      tags:
        - Admin Auth
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                  description: Full name of the admin
                  example: "Adminov Admin"
                email:
                  type: string
                  format: email
                  description: Email address
                  example: "admin@example.com"
                phone:
                  type: string
                  description: Phone number
                  example: "+998901234567"
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Profile updated successfully"
                  admin:
                    type: object
                    properties:
                      id:
                        type: string
                      username:
                        type: string
                      email:
                        type: string
                      full_name:
                        type: string
                      phone:
                        type: string
                      is_active:
                        type: boolean
        '400':
          description: Validation error or email already in use
        '401':
          description: Unauthorized
        '404':
          description: Admin not found
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT